requirements.txt
python-nmap
requests==2.26.0
colorama==0.4.4

